Chore: change directory
