--- src/main/java/alarm_sevice/domain/alarm/dto/CreateRequestDto.java
@@ -1,12 +1,0 @@
-package alarm_sevice.domain.alarm.dto;
-
-import alarm_sevice.domain.alarm.enums.ServiceType;
-
-public record CreateRequestDto(
-    Long userId,
-    ServiceType serviceType,
-    Long serviceId,
-    String title,
-    String content
-) {
-}
