+++ src/main/java/alarm_sevice/kafka/booking/RestaurantBookCancelDto.java
@@ -1,16 +1,13 @@
-package alarm_sevice.domain.alarm.kafkaDto.booking;
+package alarm_sevice.kafka.booking;
 
 import alarm_sevice.domain.alarm.enums.ServiceType;
 
-import java.time.LocalDateTime;
-
-public record RestaurantBookConfirmDto(
+public record RestaurantBookCancelDto(
         Long userId,
         ServiceType serviceType,
         Long serviceId,
         String restaurantName,
         String userName,
-        Integer people,
-        LocalDateTime date
+        String reason
 ) {
 }
