+++ src/main/java/alarm_sevice/kafka/register/ServiceRegisterRequestDto.java
@@ -1,4 +1,4 @@
-package alarm_sevice.domain.alarm.kafkaDto.register;
+package alarm_sevice.kafka.register;
 
 import alarm_sevice.domain.alarm.enums.ServiceType;
