+++ src/main/java/alarm_sevice/kafka/waiting/CustomerFromSellerCancelDto.java
@@ -1,4 +1,4 @@
-package alarm_sevice.domain.alarm.kafkaDto.waiting;
+package alarm_sevice.kafka.waiting;
 
 import alarm_sevice.domain.alarm.enums.ServiceType;
