--- src/test/java/alarm_sevice/AlarmApplicationTests.java
@@ -1,13 +1,0 @@
-package alarm_sevice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AlarmApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
